api/requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.8.0
billiard==3.3.0.23
boto3==1.9.243
botocore==1.12.243
cairocffi==1.1.0
CairoSVG==2.4.2
celery==3.1.26.post2
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-anymail==7.0.0
django-background-tasks==1.2.5
django-celery==3.3.1
django-celery-beat==2.0.0
django-compat==1.0.15
django-cors-headers==3.1.1
django-crontab==0.7.1
django-extra-fields==2.0.5
django-filter==2.2.0
django-mail-templated==2.6.5
django-phonenumber-field==4.0.0
django-render-block==0.6
django-rest-auth==0.9.5
django-simple-history==2.8.0
django-storages==1.7.2
django-templated-email==2.3.0
django-timezone-field==4.0
django-weasyprint==1.0.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
drf-extensions==0.5.0
gunicorn==19.9.0
html5lib==1.1
idna==2.8
importlib-metadata==1.6.0
jmespath==0.9.4
kombu==3.0.37
numpy==1.19.4
oauthlib==3.1.0
onesignal-notifications==0.2.6
onesignal-sdk==1.1.0
pandas==1.1.4
pexpect==4.8.0
phonenumbers==8.12.6
Pillow==8.0.1
psycopg2==2.8.5
ptyprocess==0.6.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
Pyphen==0.9.5
python-crontab==2.4.1
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.2
redis==3.5.0
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
sendgrid==6.1.0
six==1.15.0
sqlparse==0.3.0
tinycss2==1.0.2
twilio==6.35.2
urllib3==1.25.6
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
whitenoise==4.1.4
xmltodict==0.12.0
XlsxWriter==1.3.6
zipp==3.1.0
